data preparation/TTS/requirements.txt
tensorflow-gpu==1.15
torch==1.8.0
nitk==3.5
subprocess32==3.5.4
webrtcvad
umap-learn
visdom
tensorboard
librosa>=0.8.0
matplotlib>=3.3.0
numpy==1.19.3; platform_system == "Windows"
numpy==1.19.4; platform_system != "Windows"
scipy>=1.0.0
tqdm
sounddevice
SoundFile
Unidecode
inflect
PyQt5
multiprocess
numba
